Depend on original `type_condition`
Since https://github.com/rails/rails/commit/eeaf9cf,
use `where!` instead of `where` for generating relation.

As a result, `type_condition` is passed `WhereClauseFactory#build` as is.
However, when nil is passed to `WhereClauseFactory#build`, it becomes
`ArgumentError`.
In order to solve this, modify to generate type condition for mti.